+++ CHANGELOG.md
@@ -7,7 +7,9 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [0.3.1.dev] - 2020-02-19
+
+Minor features and improvements.
 
 ### Added